Refactor: organize presenter package composables
